Refactor fvtt-status to follow entrypoint+common pattern
