+++ package.json
@@ -15,6 +15,8 @@
     "./jsconfig.json": "./jsconfig.json",
     "./*": "./*",
     "./scripts/*": "./scripts/*",
+    "./scripts/common/*": "./scripts/common/*",
+    "./scripts/entrypoint/*": "./scripts/entrypoint/*",
     "./helpers/*": "./helpers/*",
     "./config/*": "./config/*",
     "./patches/*": "./patches/*",
@@ -36,6 +38,8 @@
     "scripts/pod-handler.sh",
     "scripts/fvtt-status.mjs",
     "scripts/validate-config.js",
+    "scripts/common/",
+    "scripts/entrypoint/",
     "helpers/",
     "patches/",
     "babel.config.cjs",
